requirements.txt
fastapi
numpy
uvicorn
tensorflow-cpu
json